requirements.txt
streamlit #front end
google-generativeai
python-dotenv
streamlit